--- src/components/search/curated/instructors/dan-abramov/dan-abramov-page.stories.mdx
@@ -1,14 +1,0 @@
-import {Meta, Story, Canvas} from '@storybook/addon-docs/blocks'
-import SearchDanAbromov from './index'
-
-<Meta title="components/search/curated/instructor/dan-abramov" />
-
-# Curated Landing Page
-
-<Canvas>
-  <Story name="Default">
-    <div className="bg-gray-100 p-5">
-      <SearchDanAbromov />
-    </div>
-  </Story>
-</Canvas>
