+++ src/components/search/instructors/dan-abramov/dan-abramov-page-data.ts
